requirements.txt
cement==3.0.4
discord
numpy
matplotlib
jinja2
pyyaml
colorlog
pyfiglet
colorama
termcolor
six
tinydb
tabulate
python-hosts
python-dotenv
#Needed in Windows
#windows-cursesNzMwNTEyMDQ5MzcyOTIxOTQ4.XwYoXQ.0UaAMaPp_T_b-iNBOguCjHZWwis
awscli
#aws-sam-cli
spotipy
twitchio
scikit-build
opencv-python
mahotas
pytesseract